added new confirmation window in fourth option when theres still stock in the market, and in the fifth option added a check to detect if theres not enough stock to sell

+++ src/projetoComercio/Main.java
@@ -11,15 +11,14 @@ public class Main {
         while (option != 6) {
             System.out.println("\n                                --- MAIN MENU ---\n");
             System.out.println("(1) List All (2) Register Product (3) Add Stock (4) Remove Product (5) Sell Product (6) Exit");
-            option = sc.nextInt();
-            sc.nextLine(); // Consumir a quebra de linha pendente
+            option = Integer.parseInt(sc.nextLine());
 
             if (option == 1) {
-            	operation.List();
+                operation.List();
             }
             else if (option == 2) {
 
-            	int productStock = 0;
+                int productStock = 0;
 
                 System.out.print("Enter the product name:");
                 String productName = sc.nextLine();
@@ -34,11 +33,17 @@ public class Main {
                         System.out.println("Enter the stock quantity");
                         productStock = Integer.parseInt(sc.nextLine());
                         operation.registerProduct(productName,productCode,productStock);
-                        System.out.println("Product added successfully");
+                        Products product = operation.locatePerCode(productCode);
+                        System.out.print((product.getName()+" added successfully."));
+                        System.out.println(" code: " + product.getCode() +", " + "stock: " + product.getStockQuantity());
+
                     }
                     else {
-                        System.out.println("Product added successfully");
                         operation.registerProduct(productName,productCode,productStock);
+                        Products product = operation.locatePerCode(productCode);
+                        System.out.print((product.getName()+" added successfully."));
+                        System.out.println(" code: " + product.getCode() +", " + "stock: " + product.getStockQuantity());
+
                     }
                 }
                 else{
@@ -46,7 +51,7 @@ public class Main {
                 }
             }
             else if (option == 3) {
-            	System.out.print("Enter the product code:");
+                System.out.print("Enter the product code:");
                 int productCode = Integer.parseInt(sc.nextLine());
 
                 System.out.print("Enter the stock quantity: ");
@@ -56,27 +61,43 @@ public class Main {
 
             }
             else if (option == 4) {
-            	System.out.print("Enter the product code: ");
+                System.out.print("Enter the product code: ");
                 int productCode = Integer.parseInt(sc.nextLine());
-                operation.deleteProduct(productCode);
+                if (operation.locatePerCode(productCode).getStockQuantity()==0) {
+                    operation.deleteProduct(productCode);
 
-
+                }
+                else {
+                    System.out.println("there's still stock of this product in the market");
+                    System.out.println("(1)remove anyway (2)exit");
+                    int optionRemove = Integer.parseInt(sc.nextLine());
+                    if (optionRemove == 1) {
+                        operation.deleteProduct(productCode);
+                    }
+                    else {
+                        System.out.println("returning");
+                    }
+                }
             }
             else if (option == 5) {
-            	System.out.print("Enter the product code: ");
+                System.out.print("Enter the product code: ");
                 int productCode = Integer.parseInt(sc.nextLine());
 
                 System.out.print("Enter the quantity to be sold: ");
                 int productStock = Integer.parseInt(sc.nextLine());
 
-                operation.removeStock(productStock, productCode);
+                if (operation.locatePerCode(productCode).getStockQuantity() - productStock < 0) {
+                    System.out.println("Not Enough Stock to sell");
+                }
+                else {
+                    operation.removeStock(productStock, productCode);
+                }
             }
             else if (option != 6) {
                 System.out.println("Invalid Option !");
-             
+
             }
         }
         sc.close();
     }
 }
-
